add JWT token handling

+++ FlaskStore/resources/item.py
@@ -1,20 +1,22 @@
-import uuid
 from flask.views import MethodView
 from flask_smorest import Blueprint,abort
+from flask_jwt_extended import jwt_required,get_jwt
+from sqlalchemy.exc import SQLAlchemyError
 from schemas import ItemSchema,ItemUpdateSchema
 from models import ItemModel
-from sqlalchemy.exc import SQLAlchemyError
 from db import db
 
 blp = Blueprint('Items',__name__,description="Items operations")
 
-@blp.route('/item/<string:item_id>')
+@blp.route('/item/<int:item_id>')
 class Item(MethodView):
+    @jwt_required()
     @blp.response(200,ItemSchema)
     def get(self,item_id):
         item = ItemModel.query.get_or_404(item_id)
         return item
-
+    
+    @jwt_required()
     @blp.arguments(ItemUpdateSchema)
     @blp.response(200,ItemSchema)
     def put(self,item_data,item_id):
@@ -28,7 +30,11 @@ class Item(MethodView):
         db.session.commit()
         return item
     
+    @jwt_required()
     def delete(self,item_id):
+        jwt = get_jwt()
+        if not jwt.get('is_admin'):
+            abort(401,message='Admin status required')
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -41,6 +47,7 @@ class ItemList(MethodView):
     def get(self):
         return ItemModel.query.all()
 
+    @jwt_required(fresh=True)
     @blp.arguments(ItemSchema)
     @blp.response(201,ItemSchema)
     def post(self,item_data):
